BACKEND/UserManagement/EditUsers/requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
Flask-SQLAlchemy==3.0.2
pymysql==1.0.3
python-dotenv==1.0.0
marshmallow==3.19.0
requests==2.31.0
PyJWT==2.7.0
